requirements.txt
asgiref==3.2.10
certifi==2021.10.8
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.6
djangorestframework==3.11.2
gunicorn==20.0.4
idna==3.3
itypes==1.2.0
Jinja2==3.0.2
jsonify==0.5
MarkupSafe==2.0.1
Pillow==8.4.0
pytz==2020.1
requests==2.26.0
sqlparse==0.3.1
uritemplate==4.1.1
urllib3==1.26.7
